shelly_docs/pyproject.toml
[project]
name = "shelly-docs"
version = "0.1.0"
description = "A TUI/CLI Knowledge Management application that extracts structured documentation items from Markdown"
readme = "README.md"
authors = [
    { name = "Chris Shella II", email = "christophershella2@gmail.com" }
]
requires-python = ">=3.14"
dependencies = [
    "mistletoe>=1.5.1",
    "ruamel-yaml>=0.19.1",
    "textual>=8.0.2",
    "typer>=0.24.1",
]

[dependency-groups]
dev = ["pytest>=9.0.2"]

[project.scripts]
shelly-docs = "shelly_docs.main:app"

[build-system]
requires = ["uv_build>=0.10.8,<0.11.0"]
build-backend = "uv_build"
